pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phishing-detector"
version = "1.0.0"
description = "Microservicio de detección de phishing con FastAPI y Gemini"
authors = [{name = "Security Team", email = "security@empresa.com"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.4.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "google-generativeai>=0.3.0",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
    "dnspython>=2.4.0",
    "whois>=0.9.27",
    "Levenshtein>=0.23.0",
    "spacy>=3.7.0",
    "html2text>=2020.1.16",
    "beautifulsoup4>=4.12.0",
    "publicsuffix2>=2.20191221",
    "structlog>=23.1.0",
    "prometheus-client>=0.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "black>=23.9.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "coverage>=7.3.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]